+++ R/rd-section.R
@@ -9,7 +9,10 @@ roxy_tag_rd.roxy_tag_section <- function(x, base_path, env) {
   title <- str_split(pieces[1], "\n")[[1]]
 
   if (length(title) > 1) {
-    roxy_tag_warning(x, "Section title spans multiple lines")
+    roxy_tag_warning(x, paste0(
+      "Section title spans multiple lines.\n",
+      "Did you forget a colon (:) at the end of the title?"
+    ))
     return()
   }