+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Auto-generated package documentation can now handle author ORCID comments
   containing full url (#1040).
+* The warning for section titles spanning multiple lines now includes the hint 
+  "Did you forget a colon (:) at the end of the title?" (@maelle, #994).
 
 * Can now document objects created with `delayedAssign()` by forcing
   evaluation at documentation time (#1041)